refactor: extract user input handler
