+++ src/index.js
@@ -1,48 +1,5 @@
-import {printCurrentDir, up, cd, ls} from "./nav/nav.mjs";
-import {OperationError} from "./error/error.mjs";
-import {add, cat, cp, mv, rm, rn} from "./file/file.mjs";
-import {os} from "./os/os.mjs";
-import {hash} from "./hash/hash.mjs";
-import {compress, decompress} from "./zip/brotli.mjs";
-
-const exit = (commandArgs) => {
-    if(commandArgs.size > 1) throw new Error()
-    process.exit(0)
-}
-
-const commandHandlers = {
-    ".exit": exit,
-    "up": up,
-    "cd": cd,
-    "ls": ls,
-    "cat": cat,
-    "add": add,
-    "rn": rn,
-    "cp": cp,
-    "mv": mv,
-    "rm": rm,
-    "os": os,
-    "hash": hash,
-    "compress": compress,
-    "decompress": decompress
-}
-
-const onUserInput = async (chunk) => {
-    const chunkStringified = chunk.toString().trim();
-    const commandArgs = chunkStringified.split(" ");
-    const command = commandArgs[0];
-    try {
-        const handler = commandHandlers[command]
-        await handler(commandArgs)
-    } catch (e) {
-        if(e instanceof OperationError) {
-            console.log("Operation failed")
-        } else{
-            console.log("Invalid input")
-        }
-    }
-    printCurrentDir();
-}
+import {printCurrentDir} from "./nav/nav.mjs";
+import {onUserInput} from "./controller/controller.mjs";
 
 const start = async () => {
     const username = process.argv.find((arg) => arg.startsWith("--username"))?.slice(11) || "Anonymous"
